+++ packages/meedu/lib/state.dart
@@ -3,3 +3,5 @@ library state;
 export 'src/notifiers/base_notifier.dart';
 export 'src/notifiers/simple_notifier.dart';
 export 'src/notifiers/state_notifier.dart';
+export 'src/persistent_state/persistent_state_mixin.dart'
+    show PersistentStateStorage, PersistentStateMixin;